Created and Tested CSV Import/Write Functionality
Added functionality to:
1. Import the Symbols dictionary and translation CSV file
2. Ability to write the file (for testing purposes)

This includes comments printed to the console for easy debugging.
Have also re-imported the CSV into Apple's 'Numbers' program to test that the columns and row formatting is preserved once it has been converted into the data model and back.

The next step is to write the code to import it into the model file.