Checkpoint - Summation Halfway Mark
Added Summation View Controller
Is not quite working yet but is most of the way there. Need to get the '+' subscript signs to appear in the right spot, and the constraints working so the text field is the right size.
Will still need to add a translation for it and then add a 'Product' version (which should be simple enough).
Also made the navigation controller the initial view controller.